% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accuracy.R
\name{Accuracy}
\alias{Accuracy}
\title{Accuracy}
\usage{
Accuracy(obs_est)
}
\arguments{
\item{obs_est}{A 2-column csv file.
The 1st column is the reference/observed data.
The 2nd column is the estimated/predicted data}
}
\value{
Absolute mean error (AME)
Square root of the mean squared error (RMSE)
Willmott's indices of agreemenet: original (d)
Modified (dmod) and refined (dref)
Pearson determination coefficient (R2).
All measures may have their corresponding confidence intervals (CIinf:CIsup) calculated.
}
\description{
Accuracy
}
