% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObsEst.R
\docType{data}
\name{ObsEst}
\alias{ObsEst}
\title{Example of the input required by the Accuracy function.}
\format{
A 2-column matrix with 1434 rows

\describe{
\item{ETP_obs}{PE data from a reference weather station}
\item{ETP_est}{PE data from the NASAPOWER project}
}
}
\source{
{Generated by the PowerSDI package using data from NASAPOWER and Agronomic Institute}
}
\usage{
ObsEst
}
\description{
Contains pairs of reference and estimated data.
}
\examples{
data(ObsEst)
}
\keyword{datasets}
