% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperatSDI.R
\name{OperatSDI}
\alias{OperatSDI}
\title{OperatSDI}
\usage{
OperatSDI(lonlat, start.date, end.date, DistPar, TS)
}
\arguments{
\item{lonlat}{A 2-column matrix.
The first column is the longitude in decinal degrees: (+) Estern hemisphere (-) Western hemisphere.
The second column is the latitude in decinal degrees: (+) Northern hemisphere (-) Southern hemisphere.}

\item{start.date}{Date at each the calculation must star (format; “DD-MM-YYYY").}

\item{end.date}{Date at each the calculation must end (format; “DD-MM-YYYY").}

\item{DistPar}{A csv file provided by the Scient.R function with the distributions parameters required for calculating the indices.}

\item{TS}{#'Time scale on the "quart.month" basis (integer values between 1 and 96).}
}
\value{
A csv file with Rainfall amounts, potential evapotranspiration amounts (EP; Hargreaves & Samani or FAO-56 Penman-Monteith),
differences between rainfall and EP, the NASA-SPI and NASA_SPEI,
and the SPI/SPEI classification (dry/wet categories) corresponding to each indices estimates.
}
\description{
OperatSDI
}
