% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotData.R
\name{PlotData}
\alias{PlotData}
\title{PlotData}
\usage{
PlotData(lon, lat, start.date, end.date)
}
\arguments{
\item{lon}{longitude in decinal degrees: (+) Estern Hemispher (-) Western Hemisphere.}

\item{lat}{latitude in decinal degrees: (+) Northern hemispher (-) Southern Hemisphere.}

\item{start.date}{date at which the indices estimates should start. Format: YYYY-MM-DD".}

\item{end.date}{date at which the indices estimates should end. Format: YYYY-MM-DD".}
}
\value{
Scatter plots of Rainfall and potential evapotranspiration accumulated at the 1-quart.month time scale.
}
\description{
PlotData
}
\examples{
PlotData(lon = -47.3, lat = -22.87, start.date = "2021-12-28", end.date = "2022-12-31")
}
