% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accuracy.R
\name{print.PowerSDI.Accuracy}
\alias{print.PowerSDI.Accuracy}
\title{Prints PowerSDI.Accuracy Objects}
\usage{
\method{print}{PowerSDI.Accuracy}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{a `PowerSDI.Accuracy` object}

\item{digits}{The number of digits to be used after the decimal when
displaying accuracy values.}

\item{...}{ignored}
}
\description{
Custom [print()] method for `PowerSDI.Accuracy` objects.
}
