% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference.R
\name{Reference}
\alias{Reference}
\title{Reference}
\usage{
Reference(ref)
}
\arguments{
\item{ref}{A csv file with the variables required for calculating the SPI and SPEI.
The number of variables depends on the potential evapotranspiration method selected by the user.
Variables required for both Hargreaves & Samani and FAO-56 Penman-Monteith methods:
1st Column (Head=YEAR): The year of each daily data
2nd Column (Head=MM): The month of each daily data (1 to 12)
3rd Column (Head=DD): The day of each daily data (1 to 28, 29, 30, or, 31)
4th Column (Head=tmed): Daily average air temperature at 2 meters above the ground (ºC)
5th Column (Head=tmax): Daily maximum air temperature at 2 meters above the ground (ºC)
6th Column (Head=tmin): Daily minimum air temperature at 2 meters above the ground (ºC)
7th Column (Head=Ra): Daily top of the atmosphere (extraterrestrial) radiation (MJ/m^2/day)
For the Hargreaves & Samani method
8th Column (Head=Rain): Daily rainfall amounts (mm)
Tip: Use the file referenceHS.csv as example.
For the FAO-56 Penman-Monteith method
8th Column (Head=Rs): Daily global horizontal irradiance (MJ/m^2/day)
9th Column (Head=W): Daily average wind speed at 2 meters above the ground (m/s)
10th Column (Head=RH): Daily average relative humidity at 2 meters above the ground (%)
11th Column (Head=Rain): Daily rainfall amounts (mm)
Tip: Use the file referencePM.csv as example.}
}
\value{
Scatter plots of Rainfall and potential evapotranspiration accumulated at the 1-quart.month time scale.
A csv file (ReferenceSDI.csv) with: Rain, Potential evapotranspiration,
The difference between rainfall and potential evapotranspiration,
SPI and SPEI calculated at the times scales selected by the user.
The user may also select the evapotranspiration estimation method:
Hargreaves & Samani or FAO-56 Penman-Monteith method
}
\description{
Reference
}
