% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientSDI.R
\name{ScientSDI}
\alias{ScientSDI}
\title{ScientSDI}
\usage{
ScientSDI(lon, lat, start.date, end.date, TS)
}
\arguments{
\item{lon}{longitude in decinal degrees: (+) Estern Hemispher (-) Western Hemisphere.}

\item{lat}{latitude in decinal degrees: (+) Northern hemispher (-) Southern Hemisphere.}

\item{start.date}{date at which the indices estimates should start. Format: DD-MM-YYYY".}

\item{end.date}{date at which the indices estimates should end. Format: DD-MM-YYYY".}

\item{TS}{Time scale on the "quart.month" basis (integer values between 1 and 96).}
}
\value{
Four csv files with data calculated at the time scale selected by the user.
DistPar.csv: The parameters of the distributions (gamma and GEV) used to calculate the indices.
Goodness.csv: The Lilliefors and Anderson-Darling tests goodness-of-fit tests.
Normality.csv:  The outcomes of the two normality checking procedures (Wu et al., 2007 and Stagge et., 2015)
ScientSDI.csv: The NASA-SPI, NASA-SPEI.HS and NASA-SPEI.PM.
This function also presents other data obtained from the NASAPOWER project, that is:
Rainfall amounts (Rain).
Potential evapotranspitations values estimated through the Hargreaves & Samani method (EPHS).
Potential evapotranspitations values estimated through the FAO-56 Penman-Monteith method (EPPM).
The difference between rainfall and potential evapotranspiration (P-EPHS and P-EPPM).
}
\description{
ScientSDI
}
