% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistPar.R, R/data.R
\docType{data}
\name{DistPar}
\alias{DistPar}
\title{Parameters for calculating the SDIs. Provided by the ScientSDI function.}
\format{
A 13-column matrix with 48 rows

\describe{
\item{lon}{longitude in decinal degrees}
\item{lat}{latitude in decinal degrees}
\item{quart.month}{The quartile of each month}
\item{alfa.rain}{Shape parameter of the gamma distribution}
\item{beta.rain}{Scale parameter of the gamma distribution}
\item{probzero.rain}{Probability of rain=0}
\item{loc.harg}{Location parameter of the GEV distribution, EP calculated by HS method}
\item{sc.harg}{Scale parameter of the GEV distribution, EP calculated by HS method}
\item{sh.harg}{Shape parameter of the GEV distribution, EP calculated by HS method}
\item{loc.pm}{Location parameter of the GEV distribution, EP calculated by PM method}
\item{sc.pm}{Scale parameter of the GEV distribution, EP calculated by PM method}
\item{sh.pm}{Shape parameter of the GEV distribution, EP calculated by PM method}
\item{TS}{Time scale at which the SDIs will be calculated}
}

A matrix with 48 rows and 13 variables:
\describe{
\item{lon}{Longitute of the weather station of Campinas-SP}
\item{lat}{Latitude of the weather station of Campinas-SP}
\item{quart.month}{The quasi-weekly period}
\item{alfa.rain}{Shape parameter of the gamma distribution}
\item{beta.rain}{Scale parameter of the gamma distribution}
\item{probzero.rain}{Probability of rain equal to zero}
\item{loc.harg}{Location parameter of the GEV distribution. PE calculated by HS method}
\item{sc.harg}{Scale parameter of the GEV distribution. PE calculated by HS method}
\item{sh.harg}{Shape parameter of the GEV distribution. PE calculated by HS method}
\item{loc.pm}{Location parameter of the GEV distribution. PE calculated by PM method}
\item{sc.pm}{Scale parameter of the GEV distribution. PE calculated by PM method}
\item{sh.pm}{Shape parameter of the GEV distribution. PE calculated by PM method}
}
}
\source{
{Generated by the ScientSDI.R function using NASAPOWER data}

Calculated using the ScientSDI function
}
\usage{
DistPar

DistPar
}
\description{
Contains parameters of the gamma and GEV distributions and the Pr(Rain=0).

This data can be provided by the ScientSDI function.
}
\examples{
data(DistPar)
data(DistPar)
}
\keyword{datasets}
