% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObsEst.R
\docType{data}
\name{ObsEst}
\alias{ObsEst}
\title{Example Data of the Input Required by the Accuracy Function}
\format{
A \code{data.frame} with 2 variables and 1434 rows.

\describe{
\item{PE_obs}{PE data from a reference weather station}
\item{PE_est}{PE data from the \acronym{NASA} \acronym{POWER} project}
}
}
\source{
Generated by the PowerSDI package using data from \acronym{NASA}
  \acronym{POWER} and Agronomic Institute.
}
\usage{
ObsEst
}
\description{
Contains pairs of reference and estimated data.
}
\examples{
data(ObsEst)
}
\keyword{datasets}
