% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperatSDI.R
\name{OperatSDI}
\alias{OperatSDI}
\title{OperatSDI}
\usage{
OperatSDI(
  lon,
  lat,
  start.date,
  end.date,
  PEMethod = "HS",
  distr = "GEV",
  parms,
  TS = 4
)
}
\arguments{
\item{lon}{longitude in decinal degrees.}

\item{lat}{latitude in decinal degrees.}

\item{start.date}{Date at each the calculation must start (“YYYY-MM-DD").}

\item{end.date}{Date at each the calculation must end (“YYYY-MM-DD").}

\item{PEMethod}{A character variable ("HS" or "PM") defining the potential evapotranspiration method. Default is "HS".}

\item{distr}{A character variable ("GEV" or "GLO") defining which distribution is used to calculate the SPEI. Default is "GEV".}

\item{parms}{Parameters required for calculating the SPI and SPEI. It is provided by the ScientSDI function (DistPar).}

\item{TS}{#'Time scale on the "quart.month" basis (integer values between 1 and 96).}
}
\value{
A data frame with Rainfall, potential evapotranspiration (PE),
difference between rainfall and PE (in millimiters), the NASA-SPI and NASA_SPEI,
and the SDI categories corresponding to each indices estimates.
}
\description{
Calculates the SPI and SPEI using a NASAPOWER data.
}
\examples{
data("DistPar")
 OperatSDI(lon = -47.3, lat = -22.67, start.date = "2023-01-31",
  end.date = "2023-07-07", parms = DistPar)
}
