% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OperatSDI.R
\name{OperatSDI}
\alias{OperatSDI}
\title{OperatSDI}
\usage{
OperatSDI(
  lon,
  lat,
  start.date,
  end.date,
  PEMethod = "HS",
  distr = "GEV",
  parms,
  TS = 4
)
}
\arguments{
\item{lon}{longitude in decinal degrees: (+) Estern Hemispher (-) Western Hemisphere.}

\item{lat}{latitude in decinal degrees: (+) Northern hemispher (-) Southern Hemisphere.}

\item{start.date}{Date at each the calculation must star (format; “YYYY-MM-DD").}

\item{end.date}{Date at each the calculation must end (format; “YYYY-MM-DD").}

\item{PEMethod}{A character variable ("HS" or "PM") defining the potential evapotranspiration method. Default is "HS".}

\item{distr}{A character variable ("GEV" or "GLO") defining which distribution is used to calculate the SPEI. Default is "GEV".}

\item{parms}{Parameters required for calculating the SPI and SPEI. It is provided by the ScientSDI function (DistPar).}

\item{TS}{#'Time scale on the "quart.month" basis (integer values between 1 and 96).}
}
\value{
A data frame with Rainfall, potential evapotranspiration (EP; Hargreaves & Samani or FAO-56 Penman-Monteith),
differences between rainfall and EP (in millimiters), the NASA-SPI and NASA_SPEI,
and the SPI/SPEI classification (dry/wet categories) corresponding to each indices estimates.
}
\description{
OperatSDI
}
\examples{
data("DistPar")
OperatMode <- OperatSDI(
  lon = -47.3, lat = -22.67, start.date = "2023-01-31",
  end.date = "2023-07-07", parms = DistPar
)
}
