% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accuracy.R
\name{plot.PowerSDI.Accuracy}
\alias{plot.PowerSDI.Accuracy}
\title{Plots PowerSDI.Accuracy Objects}
\usage{
\method{plot}{PowerSDI.Accuracy}(x, ...)
}
\arguments{
\item{x}{a `PowerSDI.Accuracy` object}

\item{...}{Other parameters as passed to \code{plot()}}
}
\value{
Nothing. Side-effect: plots graphs.
}
\description{
Custom \code{plot()} method for \code{PowerSDI.Accuracy} objects.
}
