% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference.R
\name{Reference}
\alias{Reference}
\title{Reference}
\usage{
Reference(ref, distr = "GEV", plotdata = "Yes", PEMethod = "HS", TS = 4)
}
\arguments{
\item{ref}{A data frame with the variables required for calculating the SDIs. See refHS or refPM as examples.}

\item{distr}{A character variable ("GEV" or "GLO") defining which distribution is used to calculate the SPEI. Default is "GEV".}

\item{plotdata}{A character variable ("Yes" or "No") defining if the function must plot the rainfall and PE data. Default is "Yes".}

\item{PEMethod}{A character variable ("HS" or "PM") defining the potential evapotranspiration method. Default is "HS".}

\item{TS}{Time scale on the quart.month" basis (integer values between 1 and 96). Default is 4.}
}
\value{
A data frame with:
Rain, potential evapotranspiration,
Difference between rainfall and potential evapotranspiration,
SPI and SPEI calculated at the time scale selected by the user.
}
\description{
Reference
}
\examples{
data("refHS")
Reference(ref = refHS, distr = "GEV", plotdata = "Yes", PEMethod = "HS", TS = 4)
}
