% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refHS.R
\docType{data}
\name{refHS}
\alias{refHS}
\title{Example of the Input Required by the Reference Function}
\format{
A \code{data.frame} with 10950 rows and 8 variables.

\describe{
\item{YEAR}{Year}
\item{MM}{Month}
\item{DD}{Day}
\item{tavg}{Daily average air temperature at 2 metres above the ground
  (degrees C)}
\item{tmax}{Daily maximum air temperature at 2 metres above the ground
  (degrees C)}
\item{tmin}{Daily minimum air temperature at 2 metres above the ground
  (degrees C)}
\item{Ra}{Daily top of the atmosphere radiation (MJ/m^2/day)}
\item{Rain}{Daily rainfall amounts (mm)}
}
}
\source{
Agronomic Institute and \acronym{NASA} \acronym{POWER}.
}
\usage{
refHS
}
\description{
Contains data for calculating the \acronym{SPI} and \acronym{SPEI}.
}
\examples{
data(refHS)
}
\keyword{datasets}
