% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refPM.R
\docType{data}
\name{refPM}
\alias{refPM}
\title{Another example of the input required by the Reference.R function.}
\format{
A 11-column matrix with 10958 rows and 11 variables
\describe{
\item{YEAR}{Year}
\item{MM}{Month}
\item{DD}{Day}
\item{tmed}{Daily average air temperature at 2 meters above the ground (ºC)}
\item{tmax}{Daily maximum air temperature at 2 meters above the ground (ºC)}
\item{tmin}{Daily minimum air temperature at 2 meters above the ground (ºC)}
\item{Ra}{Daily top of the atmosphere radiation (MJ/m^2/day)}
\item{Rs}{Daily global horizontal irradiance (MJ/m^2/day)}
\item{W}{Daily average wind speed at 2 meters above the ground (m/s)}
\item{RH}{Daily average relative humidity at 2 meters above the ground (in percentage)}
\item{Rain}{Daily rainfall amounts (mm)}
}
}
\source{
{Agronomic Institute and NASAPOWER}
}
\usage{
refPM
}
\description{
Contains data for calculating the SPI and SPEI.
}
\examples{
data(refPM)
}
\keyword{datasets}
