% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScientSDI.R
\name{ScientSDI}
\alias{ScientSDI}
\title{ScientSDI}
\usage{
ScientSDI(
  lon,
  lat,
  start.date,
  end.date,
  distr = "GEV",
  TS = 4,
  Good = "Yes",
  sig.level = 0.95,
  RainUplim = NULL,
  RainLowlim = NULL,
  PEUplim = NULL,
  PELowlim = NULL
)
}
\arguments{
\item{lon}{longitude in decinal degrees: (+) Estern Hemispher (-) Western Hemisphere.}

\item{lat}{latitude in decinal degrees: (+) Northern hemispher (-) Southern Hemisphere.}

\item{start.date}{date at which the indices estimates should start. Format: YYYY-MM-DD".}

\item{end.date}{date at which the indices estimates should end. Format: YYYY-MM-DD".}

\item{distr}{A character variable ("GEV" or "GLO") defining which distribution is used to calculate the SPEI. Default is "GEV".}

\item{TS}{Time scale on the "quart.month" basis (integer values between 1 and 96). Default is 4.}

\item{Good}{A character variable ("Yes" or "No") defining if the goodness-of-fit and normality tests. Default is "Yes".}

\item{RainUplim}{Optional. Upper limit in millimeters from which rainfall values larger than it will be removed. Default is NULL.}

\item{RainLowlim}{Optional. Lower limit in millimeters from which rainfall values smaller than it will be removed. Default is NULL.}

\item{PEUplim}{Optional. Upper limit in millimeters from which evapotranspiration values larger than it will be removed. Default is NULL.}

\item{PELowlim}{Optional. Lower limit in millimeters from which evapotranspiration values smaller than it will be removed. Default is NULL.}

\item{sig, level}{A numeric variable (between 0.90 and 0.95) defining the significance level for parameter Good. Default is "0.95".}
}
\value{
Four csv files with data calculated at the time scale selected by the user.
DistPar.csv: The parameters of the distributions (gamma and GEV) used to calculate the indices.
Goodness.csv: The Lilliefors and Anderson-Darling tests goodness-of-fit tests.
Normality.csv:  The outcomes of the two normality checking procedures (Wu et al., 2007 and Stagge et., 2015).
ScientSDI.csv: The NASA-SPI, NASA-SPEI.HS and NASA-SPEI.PM.
This function also presents other data obtained from the NASAPOWER project, that is:
Rainfall amounts (Rain).
Potential evapotranspitations values estimated through the Hargreaves & Samani method (EPHS).
Potential evapotranspitations values estimated through the FAO-56 Penman-Monteith method (EPPM).
The difference between rainfall and potential evapotranspiration (P-EPHS and P-EPPM).
}
\description{
ScientSDI
}
\examples{
ScientSDI(lon=-47.3, lat=-22.87, start.date="1991-01-01", end.date="2022-12-31",TS=4,Good="no",sig.level=0.95,distr="GEV",RainUplim=NULL,RainLowlim=NULL,PEUplim=NULL,PELowlim=NULL)
}
